Refactor: Apply UpperCamelCase in file route names
