+++ src/routes/Routes.tsx
@@ -1,5 +1,5 @@
 import { createBrowserRouter } from 'react-router-dom';
 import { AuthRoutes } from './uthRoutes';
-import { RootRoutes } from './ootRoutes';
+import { RootRoutes } from './RootRoutes';
 
 export const Routes = createBrowserRouter([AuthRoutes, RootRoutes]);
